Moves facade to its folder
